chore(runtime): move dormant systems out of core
Phase 1 of ForgeFlow diet:
- remove unused evolution_* compatibility facades
- move orchestra and experiment modules under experimental/
- update runtime imports and focused tests
- document the broader diet plan

Verification:
- 145 focused runtime tests pass
- structure validation passes
