+++ experimental/orchestra/debate.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 from typing import Any
 
 from forgeflow_runtime.executor import RunTaskRequest, RunTaskResult, dispatch
-from forgeflow_runtime.orchestra.strategy import (
+from experimental.orchestra.strategy import (
     OrchestrationConfig,
     OrchestrationResult,
     _merge_artifacts,
