+++ experimental/orchestra/fastest.py
@@ -6,7 +6,7 @@ import logging
 from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 
 from forgeflow_runtime.executor import RunTaskRequest, RunTaskResult, dispatch
-from forgeflow_runtime.orchestra.strategy import (
+from experimental.orchestra.strategy import (
     OrchestrationConfig,
     OrchestrationResult,
     _merge_artifacts,
