+++ experimental/orchestra/pipeline.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 
 from forgeflow_runtime.executor import RunTaskRequest, RunTaskResult, dispatch
-from forgeflow_runtime.orchestra.strategy import (
+from experimental.orchestra.strategy import (
     OrchestrationConfig,
     OrchestrationResult,
     _merge_token_usage,
