--- forgeflow_runtime/evolution_audit.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.audit``.
-"""
-
-from forgeflow_runtime.evolution.audit import *  # noqa: F401,F403
