--- forgeflow_runtime/evolution_cases.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.cases``.
-"""
-
-from forgeflow_runtime.evolution.cases import *  # noqa: F401,F403
