--- forgeflow_runtime/evolution_doctor.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.doctor``.
-"""
-
-from forgeflow_runtime.evolution.doctor import *  # noqa: F401,F403
