--- forgeflow_runtime/evolution_execution.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.execution``.
-"""
-
-from forgeflow_runtime.evolution.execution import *  # noqa: F401,F403
