--- forgeflow_runtime/evolution_lifecycle.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.lifecycle``.
-"""
-
-from forgeflow_runtime.evolution.lifecycle import *  # noqa: F401,F403
