--- forgeflow_runtime/evolution_observations.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.observations``.
-"""
-
-from forgeflow_runtime.evolution.observations import *  # noqa: F401,F403
