--- forgeflow_runtime/evolution_promotion_gates.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.promotion_gates``.
-"""
-
-from forgeflow_runtime.evolution.promotion_gates import *  # noqa: F401,F403
