--- forgeflow_runtime/evolution_promotion_plans.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.promotion_plans``.
-"""
-
-from forgeflow_runtime.evolution.promotion_plans import *  # noqa: F401,F403
