--- forgeflow_runtime/evolution_promotions.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.promotions``.
-"""
-
-from forgeflow_runtime.evolution.promotions import *  # noqa: F401,F403
