--- forgeflow_runtime/evolution_proposals.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.proposals``.
-"""
-
-from forgeflow_runtime.evolution.proposals import *  # noqa: F401,F403
