--- forgeflow_runtime/evolution_rules.py
@@ -1,6 +1,0 @@
-"""Compatibility shim for the evolution subpackage.
-
-New code should import from ``forgeflow_runtime.evolution.rules``.
-"""
-
-from forgeflow_runtime.evolution.rules import *  # noqa: F401,F403
