+++ tests/runtime/test_experiment.py
@@ -10,16 +10,16 @@ from typing import Any
 
 import pytest
 
-from forgeflow_runtime.experiment.circuit import CircuitBreaker, CircuitState
-from forgeflow_runtime.experiment.git_ops import ExperimentGit, GitDiff
-from forgeflow_runtime.experiment.loop import (
+from experimental.experiment.circuit import CircuitBreaker, CircuitState
+from experimental.experiment.git_ops import ExperimentGit, GitDiff
+from experimental.experiment.loop import (
     ExperimentConfig,
     ExperimentResult,
     IterationResult,
     run_experiment,
 )
-from forgeflow_runtime.experiment.metric import MetricResult, execute_metric, extract_json_values
-from forgeflow_runtime.experiment.simplicity import improvement_efficiency, simplicity_score
+from experimental.experiment.metric import MetricResult, execute_metric, extract_json_values
+from experimental.experiment.simplicity import improvement_efficiency, simplicity_score
 
 
 # ---------------------------------------------------------------------------
